# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prime Home LB Ecommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount?: number;
  createdAt?: string;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  stock: number;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  images?: string[];
  featured?: boolean;
  bestSeller?: boolean;
  onSale?: boolean;
  visible?: boolean;
  salesCount?: number;
  createdAt?: string;
}

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  salePrice?: number;
  stock: number;
  sku?: string;
  categoryId?: number;
  images?: string[];
  featured?: boolean;
  bestSeller?: boolean;
  onSale?: boolean;
  visible?: boolean;
}

export interface ProductUpdate {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  /** @nullable */
  salePrice?: number | null;
  stock?: number;
  sku?: string;
  /** @nullable */
  categoryId?: number | null;
  images?: string[];
  featured?: boolean;
  bestSeller?: boolean;
  onSale?: boolean;
  visible?: boolean;
}

export interface ProductList {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface FeaturedProducts {
  featured: Product[];
  bestSellers: Product[];
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  banned?: boolean;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
}

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  quantity: number;
  imageUrl: string;
  stock?: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  deliveryFee?: number;
  discount?: number;
  total: number;
  /** @nullable */
  couponCode?: string | null;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  /** @nullable */
  productImageUrl?: string | null;
  price: number;
  quantity: number;
  subtotal?: number;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  customerName?: string;
  /** @nullable */
  customerEmail?: string | null;
  customerPhone?: string;
  address?: string;
  city?: string;
  /** @nullable */
  deliveryArea?: string | null;
  /** @nullable */
  notes?: string | null;
  subtotal?: number;
  deliveryFee?: number;
  discount?: number;
  total: number;
  paymentMethod?: string;
  /** @nullable */
  couponCode?: string | null;
  items?: OrderItem[];
  createdAt: string;
}

export interface OrderList {
  items: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface OrderInput {
  customerName: string;
  customerEmail?: string;
  customerPhone: string;
  address: string;
  city: string;
  deliveryArea?: string;
  notes?: string;
  paymentMethod: string;
  couponCode?: string;
  items: CartItemInput[];
}

export interface OrderStatusUpdate {
  status: string;
}

export interface WishlistStatus {
  inWishlist: boolean;
}

export interface TopProduct {
  productId: number;
  productName: string;
  /** @nullable */
  imageUrl?: string | null;
  salesCount: number;
  revenue: number;
}

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalCustomers: number;
  totalProducts: number;
  pendingOrders?: number;
  lowStockProducts?: number;
  revenueToday?: number;
  ordersToday?: number;
  topProducts?: TopProduct[];
  recentOrders?: Order[];
}

export interface SalesDataPoint {
  label: string;
  revenue: number;
  orders: number;
}

export interface Customer {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  orderCount: number;
  totalSpent: number;
  banned?: boolean;
  createdAt: string;
}

export interface CustomerList {
  items: Customer[];
  total: number;
}

export interface CustomerUpdate {
  banned?: boolean;
}

export interface Coupon {
  id: number;
  code: string;
  type: string;
  value: number;
  /** @nullable */
  minOrderAmount?: number | null;
  /** @nullable */
  maxUses?: number | null;
  usedCount?: number;
  /** @nullable */
  expiresAt?: string | null;
  active: boolean;
  createdAt?: string;
}

export type CouponInputType = typeof CouponInputType[keyof typeof CouponInputType];


export const CouponInputType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface CouponInput {
  code: string;
  type: CouponInputType;
  value: number;
  minOrderAmount?: number;
  maxUses?: number;
  expiresAt?: string;
}

export interface CouponValidateInput {
  code: string;
  orderAmount: number;
}

export interface CouponValidateResult {
  valid: boolean;
  discount: number;
  /** @nullable */
  message?: string | null;
}

export interface StoreSettings {
  storeName?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  deliveryFee?: number;
  /** @nullable */
  freeDeliveryThreshold?: number | null;
  currency?: string;
  /** @nullable */
  announcementText?: string | null;
  announcementActive?: boolean;
}

export interface StoreSettingsUpdate {
  storeName?: string;
  logoUrl?: string;
  whatsappNumber?: string;
  instagramUrl?: string;
  email?: string;
  address?: string;
  deliveryFee?: number;
  freeDeliveryThreshold?: number;
  currency?: string;
  announcementText?: string;
  announcementActive?: boolean;
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
minPrice?: number;
maxPrice?: number;
sort?: ListProductsSort;
page?: number;
limit?: number;
featured?: boolean;
onSale?: boolean;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  popular: 'popular',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
} as const;

export type ListOrdersParams = {
status?: string;
page?: number;
limit?: number;
search?: string;
};

export type GetSalesChartParams = {
period?: GetSalesChartPeriod;
};

export type GetSalesChartPeriod = typeof GetSalesChartPeriod[keyof typeof GetSalesChartPeriod];


export const GetSalesChartPeriod = {
  daily: 'daily',
  monthly: 'monthly',
} as const;

export type ListCustomersParams = {
page?: number;
search?: string;
};