# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Prime Home LB Ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me/update:
    patch:
      operationId: updateMe
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ─── CART ────────────────────────────────────────────────────────────────────
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      responses:
        "204":
          description: Cart cleared
  /cart/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ─── ORDERS ──────────────────────────────────────────────────────────────────
...
  /admin/stats/sales:
    get:
      operationId: getSalesChart
      tags: [admin]
      summary: Sales chart data (daily/monthly)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, monthly]
      responses:
        "200":
          description: Sales data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesDataPoint"
  /admin/customers:
    get:
      operationId: listCustomers
      tags: [admin]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
  /admin/customers/{id}:
    get:
      operationId: getCustomer
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ─── COUPONS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        phone: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        role: { type: string }
        banned: { type: boolean }
        createdAt: { type: string }
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }

    # ─── CART ────────────────────────────────────────────────────────────────
    CartItem:
      type: object
      required: [productId, productName, price, quantity, imageUrl]
      properties:
        productId: { type: integer }
        productName: { type: string }
        price: { type: number }
        salePrice: { type: ["number", "null"] }
        quantity: { type: integer }
        imageUrl: { type: string }
        stock: { type: integer }
    Cart:
      type: object
      required: [items, subtotal, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal: { type: number }
        deliveryFee: { type: number }
        discount: { type: number }
        total: { type: number }
        couponCode: { type: ["string", "null"] }
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer }

    # ─── ORDER ────────────────────────────────────────────────────────────────
...
    SalesDataPoint:
      type: object
      required: [label, revenue, orders]
      properties:
        label: { type: string }
        revenue: { type: number }
        orders: { type: integer }
    Customer:
      type: object
      required: [id, name, email, orderCount, totalSpent, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        orderCount: { type: integer }
        totalSpent: { type: number }
        banned: { type: boolean }
        createdAt: { type: string }
    CustomerList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total: { type: integer }
    CustomerUpdate:
      type: object
      properties:
        banned: { type: boolean }

    # ─── COUPONS ─────────────────────────────────────────────────────────────
...