# lib/api-zod/src/generated/types/customerList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Prime Home LB Ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from './customer';

export interface CustomerList {
  items: Customer[];
  total: number;
}